Делит код на блоки
